+++ packages/eas-cli/src/graphql/queries/DeviceRunSessionQuery.ts
@@ -54,6 +54,7 @@ export const DeviceRunSessionQuery = {
               deviceRunSessions {
                 byId(deviceRunSessionId: $deviceRunSessionId) {
                   id
+                  name
                   status
                   type
                   platform
@@ -142,6 +143,7 @@ export const DeviceRunSessionQuery = {
                       cursor
                       node {
                         id
+                        name
                         status
                         type
                         platform